+++ src/main/java/com/neo/dao/PubliciteDAO.java
@@ -9,6 +9,6 @@ public interface PubliciteDAO {
 	public void creer(Publicite publicite);
 	public void modifier(Publicite publicite);
 	public void supprimer(Publicite publicite);
-	public Publicite findById(int id);
+	public Publicite findById(long id);
 	public List<Publicite> lister();
 }
